Simplification of the UCERF calculators
